requirements.txt
python-slugify==5.0.2
text-unidecode==1.3
